Use interpreter_backend as a static library
