+++ docs/template_plugin/backend/CMakeLists.txt
@@ -1,0 +1,65 @@
+# Copyright (C) 2018-2021 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
+#
+
+set (SRC
+    backend.cpp
+    backend.hpp
+    cache.cpp
+    cache.hpp
+    executable.cpp
+    executable.hpp
+    performance_counter.hpp
+    pass/dyn_elimination.cpp
+    pass/dyn_elimination.hpp
+    pass/shape_relevance.cpp
+    pass/shape_relevance.hpp
+    int_backend.cpp int_executable.cpp evaluates_map.cpp
+    )
+
+add_library(interpreter_backend STATIC EXCLUDE_FROM_ALL ${SRC})
+
+ie_faster_build(interpreter_backend UNITY)
+
+ie_add_vs_version_file(NAME interpreter_backend
+                       FILEDESCRIPTION "OpenVINO interpreter backend")
+
+ie_add_api_validator_post_build_step(TARGET interpreter_backend)
+
+target_compile_definitions(interpreter_backend
+    PRIVATE
+        SHARED_LIB_PREFIX="${CMAKE_SHARED_LIBRARY_PREFIX}"
+        SHARED_LIB_SUFFIX="${IE_BUILD_POSTFIX}${CMAKE_SHARED_LIBRARY_SUFFIX}"
+)
+target_link_libraries(interpreter_backend PRIVATE ngraph::builder ngraph::reference
+                                     pugixml::static ov_shape_inference openvino::util openvino::itt)
+
+target_include_directories(interpreter_backend PUBLIC $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}>)
+
+install(TARGETS interpreter_backend
+        RUNTIME DESTINATION ${IE_CPACK_RUNTIME_PATH} COMPONENT tests OPTIONAL EXCLUDE_FROM_ALL
+        LIBRARY DESTINATION ${IE_CPACK_LIBRARY_PATH} COMPONENT tests OPTIONAL EXCLUDE_FROM_ALL)
+if(NOT BUILD_SHARED_LIBS)
+    install(TARGETS interpreter_backend
+            RUNTIME DESTINATION tests COMPONENT tests OPTIONAL EXCLUDE_FROM_ALL
+            LIBRARY DESTINATION tests COMPONENT tests OPTIONAL EXCLUDE_FROM_ALL)
+endif()
+
+file(GLOB_RECURSE all_backends_src "${CMAKE_CURRENT_SOURCE_DIR}/*.cpp" "${CMAKE_CURRENT_SOURCE_DIR}/*.hpp")
+add_clang_format_target(interpreter_backend_clang FOR_SOURCES ${all_backends_src})
+
+# ie_add_api_validator_post_build_step(TARGET interpreter_backend)
+# 
+# target_compile_definitions(interpreter_backend PRIVATE INTERPRETER_BACKEND_EXPORTS)
+# target_link_libraries(interpreter_backend PUBLIC ngraph_backend
+#                                           PRIVATE pugixml::static ov_shape_inference ngraph_s
+#                                                   openvino::util ngraph_reference)
+# 
+# install(TARGETS interpreter_backend
+#         RUNTIME DESTINATION ${IE_CPACK_RUNTIME_PATH} COMPONENT tests OPTIONAL EXCLUDE_FROM_ALL
+#         LIBRARY DESTINATION ${IE_CPACK_LIBRARY_PATH} COMPONENT tests OPTIONAL EXCLUDE_FROM_ALL)
+# if(NOT BUILD_SHARED_LIBS)
+#     install(TARGETS interpreter_backend
+#             RUNTIME DESTINATION tests COMPONENT tests OPTIONAL EXCLUDE_FROM_ALL
+#             LIBRARY DESTINATION tests COMPONENT tests OPTIONAL EXCLUDE_FROM_ALL)
+# endif()
