+++ docs/template_plugin/backend/cache.hpp
