+++ docs/template_plugin/backend/evaluates_map.hpp
@@ -3,7 +3,6 @@
 //
 
 #pragma once
-#include "int_backend_visibility.hpp"
 #include "ngraph/node.hpp"
 
 namespace ngraph {
