+++ docs/template_plugin/backend/pass/shape_relevance.hpp
@@ -4,12 +4,11 @@
 
 #pragma once
 
-#include "backend_visibility.hpp"
 #include "ngraph/pass/pass.hpp"
 
 namespace ngraph {
 namespace pass {
-class BACKEND_API ShapeRelevance : public FunctionPass {
+class ShapeRelevance : public FunctionPass {
 public:
     ShapeRelevance() : FunctionPass() {}
     bool run_on_model(const std::shared_ptr<ngraph::Function>& m) override;
