+++ src/core/tests/main.cpp
@@ -8,8 +8,6 @@
 #include "gtest/gtest.h"
 #include "ngraph/log.hpp"
 #include "ngraph/ngraph.hpp"
-#include "runtime/backend.hpp"
-#include "runtime/backend_manager.hpp"
 
 using namespace std;
 
@@ -30,8 +28,6 @@ int main(int argc, char** argv) {
             cpath = argv[i];
         }
     }
-    ngraph::runtime::Backend::set_backend_shared_library_search_directory(cpath);
-
     int rc = RUN_ALL_TESTS();
 
     return rc;
