--- src/core/tests/runtime/interpreter/unit_test.manifest
@@ -1,127 +1,0 @@
-INTERPRETER.onnx_top_k_opset_10
-
-# Temporarily disabled:
-INTERPRETER.onnx_resize11_scales_nearest_asymmetric_floor_dynamic_sizes
-INTERPRETER.onnx_resize11_up_sizes_cubic_half_pixel_dynamic_sizes
-INTERPRETER.interpolate_down_scales_const_linear
-INTERPRETER.onnx_resize10_up_scales_const_nearest
-INTERPRETER.onnx_resize10_up_scales_const_linear
-INTERPRETER.onnx_resize10_down_scales_const_nearest
-
-# Failed in MacOS:
-INTERPRETER.onnx_resize11_sizes_nearest_asymmetric_floor
-
-# Disabled tests for disabled reference implementations
-INTERPRETER.onnx_dyn_shapes_expand_uint16_dyn_shape
-INTERPRETER.sum_2d_to_scalar_int8
-INTERPRETER.product_to_scalar_int8
-INTERPRETER.max_pool_uint8
-INTERPRETER.max_pool_int8
-INTERPRETER.numeric_double_nan
-INTERPRETER.numeric_double_inf
-INTERPRETER.min_to_scalar_int8
-INTERPRETER.max_trivial_int8
-INTERPRETER.max_to_scalar_int8
-INTERPRETER.max_3d_to_scalar_double
-INTERPRETER.gather_4d_indices_axis_0_uint8
-INTERPRETER.gather_axis_0_int8
-INTERPRETER.gather_axis_0_int16
-INTERPRETER.gather_axis_0_uint8
-INTERPRETER.gather_axis_0_uint16
-INTERPRETER.gather_v7_4d_indices_axis_0_uint8
-INTERPRETER.gather_v7_axis_0_int8
-INTERPRETER.gather_v7_axis_0_int16
-INTERPRETER.gather_v7_axis_0_uint8
-INTERPRETER.gather_v7_axis_0_uint16
-INTERPRETER.auto_bcast_binary_elementwise
-INTERPRETER.auto_bcast_binary_elementwise_pdpd
-
-# Revise reference implementation
-onnx_dyn_model_hardmax
-onnx_model_quantize_linear_const_scale_const_zero_p
-onnx_model_quantize_linear
-onnx_model_quantize_linear_axis_zero
-onnx_model_quantize_linear_axis_negative
-
-# Backward conv
-INTERPRETER.convolution_2d_1item
-INTERPRETER.convolution_2d_1item_padded_1_1x1_1
-INTERPRETER.convolution_2d_1item_padded_2_3x4_5
-INTERPRETER.convolution_2d_2items
-INTERPRETER.convolution_2d_2items_strided
-INTERPRETER.convolution_2d_2items_strided_padded
-INTERPRETER.convolution_2d_2items_strided_padded_same
-INTERPRETER.convolution_2d_2items_dilated
-INTERPRETER.convolution_2d_2items_dilated_padded
-INTERPRETER.convolution_3d_2items
-INTERPRETER.convolution_2d_1item_1o1i_data_dilated
-INTERPRETER.convolution_2d_1item_2o1i_data_dilated
-INTERPRETER.convolution_2d_1item_2o2i_data_dilated
-INTERPRETER.convolution_2d_1item_5o3i_data_dilated
-INTERPRETER.convolution_2d_2item_5o3i_data_dilated
-
-# Removed opset0 operations
-INTERPRETER.onnx_model_dequantize_linear
-INTERPRETER.onnx_model_dequantize_linear_scalar_zero_scale_uint8
-INTERPRETER.onnx_model_dequantize_linear_scalar_zero_scale_int8
-INTERPRETER.onnx_model_dequantize_linear_1d_zero_scale_uint8
-INTERPRETER.onnx_model_dequantize_linear_1d_zero_scale_int8
-INTERPRETER.onnx_model_dequantize_linear_1d_zero_scale_int8_4d
-INTERPRETER.onnx_model_dequantize_linear_1d_zero_scale_uint8_negative_axis
-
-# If doesn't have evaluate method and we cannot infer it
-INTERPRETER.onnx_if_inside_if
-INTERPRETER.onnx_if_inside_loop
-
-# Legacy tests with unsupported features from opset4 LSTM/GRU/RNN
-# Peepholes input unsupported
-onnx_model_lstm_fwd_with_clip_peepholes
-onnx_model_lstm_bdir_short_input_seq_peepholes
-# Activation function hardsigmoid unsupported
-onnx_model_gru_fwd_activations_relu_hardsigmoid
-onnx_model_lstm_fwd_hardsigmoid_activation
-gru_cell_hardsigmoid_activation_function
-
-# Check 'n_data_channels % groups == 0' failed
-dyn_group_convolution_backprop_data
-
-# Check 'num_dyn_nodes_this_pass < num_dyn_nodes_last_pass' failed
-dyn_convolution_backprop_data
-
-# unsupported element type f16
-INTERPRETER.ctc_greedy_decoder_f16
-
-# Issue 37473. Fails on ia32 platforms only
-onnx_model_softmax_axis_0
-onnx_model_reshape_negative_dim
-
-# LogSoftmax's reference implementation doesn't handle scalar input properly
-onnx_model_logsoftmax_0D
-# Input body shape is changed during Loop iterations
-# Exception is throw during Loop shape inference
-onnx_controlflow_loop_concat_values
-
-# Infinitive Loop is not supported
-onnx_controlflow_loop_infinite
-
-# Dynamic shape support
-onnx_controlflow_loop_2d_trip_count_dynamic
-onnx_controlflow_loop_no_variadic_inputs_and_outputs
-onnx_controlflow_loop_power
-squeeze_dynamic
-
-# The test fails in CI on Ubuntu i386
-# There's an overflow of some kind: 2147483647 is not close to -2147483648 at index 2
-quantize_clamp_int32
-
-# u16 type is not supported in Minimum evaluate method
-minimum_u16
-
-# Interpreter backend doesn't implement evaluate method for OP ExperimentalDetectronROIFeatureExtractor
-# INTERPRETER.onnx_model_experimental_detectron_roi_feature_extractor
-
-# No evaluator for DeformableConv2D
-onnx_model_deformable_conv_2d
-
-# No support for unsigned types
-INTERPRETER.zero_sized_negative
